+++ Scripts/Popups/GameBoard/BoardCardEditorPopup.cs
@@ -10,7 +10,7 @@ namespace DebugMenu.Scripts.Popups.DeckEditorPopup;
 
 public class BoardCardEditorPopup : BaseWindow
 {
-    public override string PopupName => "Card Editor";
+    public override string PopupName => "Board Card Editor";
     public override Vector2 Size => new(600f, 768f);
 
     public PlayableCard currentSelection = null;
@@ -18,10 +18,9 @@ public class BoardCardEditorPopup : BaseWindow
     public override void OnGUI()
     {
         base.OnGUI();
-
         if (currentSelection == null)
         {
-            GUILayout.Label("No card selected.");
+            GUILayout.Label("No card selected!", Helpers.HeaderLabelStyle());
             return;
         }